Added installation of base kvalobs packages
Summary: This provides installation of the kvalobs package, along with configuration and starting of the four programs that provides kvalobs' core functionality. Note that configuration of teh database is still not complete, so while the programs will start, they can still not connect to a database.

Test Plan:
$ kitchen converge
$ kitchen verify

Reviewers: #it-geo-tf, #kvalobs, borgem

Reviewed By: #kvalobs, borgem

Subscribers: borgem

Maniphest Tasks: T1699

Differential Revision: https://phab.met.no/D908